Load module documentation
Previously, we could only print documentation for the items listed in
all.html.  With this patch, we also load the documentation for modules
(including crates) from the <module>/index.html file.  Note that items
are preferred over modules if there is documentation for both (e. g.
std::u8 is both a module and a primitive).