Fix path for search index items
rustdoc clears the path of an item in the search index if it is equal to
the path of the previous item (see build_index in html/render.rs).
Previously, we used the crate name if the path was empty.  With this
patch, we use the past of the last item instead.