Store full name in doc::Doc
This patch adds a new field, name, to the doc::Doc struct for
documentation items.  It stores the fully qualified name of the
documented element.  Since we no longer require a title to identify an
item, we make the title optional.  If it is not set, we print the item
name instead.