Extract parse_module_doc from parse_item_doc
Previously, we assumed that we can parse the documentation for a module
and an item (struct, trrait, …) with the same procedure.  This approach
was too simple.  For example, functions in a module are listed in a
table while the methods of a struct are listed in divs grouped by
subheadings.

Therefore, we extract the parsing of module documentation into a new
function parse_module_doc.