Use only direct childs of #main for description
Some elements, for example std::io::ErrorKind, have nested docblocks in
their type-decl docblock.  This would lead to a wrong description
because we pick the first non-type-decl docblock for the description.

With this patch, we require that the description docblock is a direct
child of #main to fix this issue.