Display member type in viewer::text::TextViewer
Previously, we only displayed the member name when viewing the
documentation of a member.  With this patch, we use the ItemType enum to
also print the type of the member.

For consistency, we remove the title field from doc::Doc that was
previously used for the titles of modules and items and use the same
title as for members.  While we lose the links to the parent items,
these were not particularly helpful anyway.  If we want to, we could
also generate them from the item name.