Add MemberGroup struct for groups of members items
Previosuly, we stored the members as a vector of vectors, group by a
string -- their title.  This was very hard to read and also too simple:
For example, there can be multiple blocks of methods in the
documentation for a struct.  Therefore we introduce a new type,
MemberGroup, that represents a group of member items with an optional
title.