Pass viewer args to viewer implementations
This patch adds the ViewerArgs struct to the Opt struct that contains
the viewer-specific arguments.  It also adds a ViewerArgs argument to
the methods of the Viewer trait so that viewers can access these
arguments.

We also refactor the TextViewer implementations so that we can pass the
viewer arguments to the Printer constructor.

Currently, there are no viewer arguments, but we will add them in
upcoming patches.