Ignore match_like_matches_macro clippy lint
This lint tells us to use the std::matches! macro instead of a match or
if let expression that yields a bool.  But the matches! macro was added
in Rust 1.42, and our MSRV is 1.40.  I don’t want to bump the MSRV only
for this change, so we ignore the lint.